・テスト表示と採点機能をひとまとめに（TestController) ・Topの遷移をloginの遷移と同じに(AuthenticationController) ・採点履歴画面に(HistoryController) ・重複した設定かつ、エラーになるため設定を追加(application.properties)
