+++ AutoSystem_springboot/src/main/java/com/example/controller/TestContoroller.java
@@ -2,6 +2,7 @@ package com.example.controller;
 
 import java.sql.Timestamp;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,18 +15,40 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import com.example.dto.ResultForm;
 import com.example.dto.TestForm;
 import com.example.entity.AnswersEntity;
+import com.example.entity.QuestionsEntity;
 import com.example.service.AnswersService;
+import com.example.service.QuestionsService;
 
 	/**
 	 * テスト Controller
 	 */
 	@Controller
-	public class AnswersContoroller {
+	public class TestContoroller {
 	/**
 	 * テスト Service
 	 */
 	@Autowired
+	private QuestionsService questionsService;
+
+	@Autowired
 	private AnswersService answersService;
+	
+	
+	/**
+	 * テストを表示
+	 * @param model Model
+	 * @return テスト
+	 */
+	@RequestMapping(value="/top", method=RequestMethod.POST,params="action=test")
+	public String postTest(Model model) {
+		//質問を取得
+		List<QuestionsEntity> questions = questionsService.searchAll();
+		//リストの中身をシャッフル
+		Collections.shuffle(questions);
+		//テスト画面へ
+		model.addAttribute("questions", questions);
+		return "test";
+	}
 
 	/**
 	 * 採点結果を表示
