+++ AutoSystem_springboot/src/main/resources/application.properties
@@ -5,12 +5,11 @@ spring.datasource.password=himitu
 # MySQLドライバー
 spring.datasource.driverClassName=com.mysql.cj.jdbc.Driver
 spring.jpa.database=MYSQL
-spring.session.store-type=jdbc
 # ポートの指定
 server.port=8084
 #セッション管理
 spring.main.allow-bean-definition-overriding=true
 spring.session.store-type=jdbc
+spring.session.jdbc.initialize-schema: always
 # タイムリーフ系のキャッシュ削除
 spring.thymeleaf.cache=false
-spring.template.cache=false
